+++ UC4-chaincode/src/main/java/de/upb/cs/uc4/chaincode/contract/ContractUtil.java
@@ -22,8 +22,13 @@ import org.hyperledger.fabric.shim.ledger.KeyValue;
 import org.hyperledger.fabric.shim.ledger.QueryResultsIterator;
 
 import java.security.NoSuchAlgorithmException;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+
+import static java.time.temporal.ChronoUnit.SECONDS;
 
 abstract public class ContractUtil {
 
@@ -103,7 +108,6 @@ abstract public class ContractUtil {
     }
 
     public GenericError getParamNumberError() {
-        // TODO add this error to operation api (approveTransaction)
         return new GenericError()
                 .type("HLParameterNumberError")
                 .title("The given number of parameters does not match the required number of parameters for the specified transaction");
@@ -123,30 +127,39 @@ abstract public class ContractUtil {
         }
 
         OperationContractUtil oUtil = new OperationContractUtil();
-        ApprovalList approvals;
         String key;
         try {
             key = OperationContractUtil.getDraftKey(contractName, transactionName, jsonArgs);
         } catch (NoSuchAlgorithmException e) {
             throw new ValidationError(GsonWrapper.toJson(getInternalError()));
         }
+        ApprovalList approvals;
+        OperationDataState operationState;
         try{
-            approvals = oUtil.getState(stub, key, OperationData.class).getExistingApprovals();
+            OperationData operation = oUtil.getState(stub, key, OperationData.class);
+            approvals = operation.getExistingApprovals();
+            operationState = operation.getState();
         } catch (Exception e) {
             approvals = new ApprovalList();
+            operationState = OperationDataState.PENDING;
         }
         String clientId = getEnrollmentIdFromClientId(ctx.getClientIdentity().getId());
         List<String> clientGroups = new GroupContractUtil().getGroupNamesForUser(ctx.getStub(), clientId);
         approvals.addUsersItem(clientId);
         approvals.addGroupsItems(clientGroups);
 
-        if(!OperationContractUtil.covers(requiredApprovals, approvals)){
+        if(operationState != OperationDataState.PENDING || !OperationContractUtil.covers(requiredApprovals, approvals)){
             throw new ValidationError(GsonWrapper.toJson(getInsufficientApprovalsError()));
         }
     }
 
     public String getEnrollmentIdFromClientId(String clientId) {
         return clientId.substring(9).split(",")[0];
+    }
+
+    public String getTimestamp(ChaincodeStub stub) {
+        DateTimeFormatter fm = DateTimeFormatter.ISO_LOCAL_DATE_TIME.withZone(ZoneId.from(ZoneOffset.UTC));
+        return fm.format(stub.getTxTimestamp().truncatedTo(SECONDS));
     }
 
     public void finishOperation(