+++ UC4-chaincode/src/test/java/de/upb/cs/uc4/chaincode/OperationContractTest.java
@@ -16,6 +16,7 @@ import org.junit.jupiter.api.function.Executable;
 import java.lang.reflect.Type;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -49,8 +50,6 @@ public final class OperationContractTest extends TestCreationBase {
                 return DynamicTest.dynamicTest(test.getName(), rejectTransactionSuccessTest(setup, input, compare, ids));
             case "rejectTransaction_FAILURE":
                 return DynamicTest.dynamicTest(test.getName(), rejectTransactionFailureTest(setup, input, compare, ids));
-            case "getOperationData_SUCCESS":
-                return DynamicTest.dynamicTest(test.getName(), getOperationDataSuccessTest(setup, input, compare, ids));
             case "finishOperation":
                 return DynamicTest.dynamicTest(test.getName(), finishOperationTest(setup, input, compare, ids));
             default:
@@ -68,12 +67,10 @@ public final class OperationContractTest extends TestCreationBase {
             Context ctx = TestUtil.mockContext(stub);
 
             for (int i = 0; i < compare.size(); i++) {
-                String operations = contract.getOperations(ctx, input.get(i*4), input.get(i*4+1), input.get(i*4+2), input.get(i*4+3));
-                Type listType = new TypeToken<ArrayList<OperationData>>() {
-
-                }.getType();
-                ArrayList<OperationData> operationDataList = GsonWrapper.fromJson(operations, listType);
-                List<String> operationIds = operationDataList.stream().map(operationData -> operationData.getOperationId()).collect(Collectors.toList());
+                String operations = contract.getOperations(ctx, input.get(i*6), input.get(i*6+1), input.get(i*6+2), input.get(i*6+3), input.get(i*6+4), input.get(i*6+5));
+                Type listType = new TypeToken<ArrayList<OperationData>>() {}.getType();
+                List<OperationData> operationDataList = GsonWrapper.fromJson(operations, listType);
+                List<String> operationIds = operationDataList.stream().map(OperationData::getOperationId).collect(Collectors.toList());
                 assertThat(GsonWrapper.toJson(operationIds)).isEqualTo(compare.get(i));
             }
 
@@ -92,7 +89,7 @@ public final class OperationContractTest extends TestCreationBase {
             for (int i = 0; i < ids.size(); i++) {
                 Context ctx = TestUtil.mockContext(stub, ids.get(i));
                 OperationData compareResult = GsonWrapper.fromJson(compare.get(i), OperationData.class);
-                OperationData transactionResult = GsonWrapper.fromJson(contract.approveTransaction(ctx, initiator(input), contract(input), transaction(input), params(input)), OperationData.class);
+                OperationData transactionResult = GsonWrapper.fromJson(contract.initiateOperation(ctx, initiator(input), contract(input), transaction(input), params(input)), OperationData.class);
                 assertThat(GsonWrapper.toJson(transactionResult)).isEqualTo(GsonWrapper.toJson(compareResult)); // TODO remove serialization
             }
         };
@@ -108,7 +105,7 @@ public final class OperationContractTest extends TestCreationBase {
             MockChaincodeStub stub = TestUtil.mockStub(setup, "UC4.OperationData:approveTransaction");
             for (String s : compare) {
                 Context ctx = cUtil.valueUnset(ids) ? TestUtil.mockContext(stub) : TestUtil.mockContext(stub, ids.get(0));
-                String result = contract.approveTransaction(ctx, "", contract(input), transaction(input), params(input));
+                String result = contract.initiateOperation(ctx, "", contract(input), transaction(input), params(input));
                 assertThat(result).isEqualTo(s);
             }
         };
@@ -129,7 +126,7 @@ public final class OperationContractTest extends TestCreationBase {
                 assertThat(GsonWrapper.toJson(transactionResult)).isEqualTo(GsonWrapper.toJson(compareResult)); // TODO remove serialization
             }*/
             Context ctx = TestUtil.mockContext(stub);
-            String rejectResult = contract.rejectTransaction(ctx, input.get(0), input.get(1));
+            String rejectResult = contract.rejectOperation(ctx, input.get(0), input.get(1));
             OperationData compareOperationData = GsonWrapper.fromJson(compare.get(compare.size() - 1), OperationData.class);
             OperationData ledgerOperationData = cUtil.getState(ctx.getStub(), compareOperationData.getOperationId(), OperationData.class);
             assertThat(GsonWrapper.toJson(compareOperationData)).isEqualTo(GsonWrapper.toJson(ledgerOperationData));
@@ -153,23 +150,8 @@ public final class OperationContractTest extends TestCreationBase {
                 assertThat(result).isEqualTo(s);
             }*/
             Context ctx = TestUtil.mockContext(stub);
-            String result = contract.rejectTransaction(ctx, input.get(0), input.get(1));
+            String result = contract.rejectOperation(ctx, input.get(0), input.get(1));
             assertThat(result).isEqualTo(compare.get(0));
-        };
-    }
-
-    private Executable getOperationDataSuccessTest(
-            JsonIOTestSetup setup,
-            List<String> input,
-            List<String> compare,
-            List<String> ids
-    ) {
-        return () -> {
-            MockChaincodeStub stub = TestUtil.mockStub(setup, "UC4.OperationData:getOperationData");
-            Context ctx = TestUtil.mockContext(stub);
-            String getResult = contract.getOperationData(ctx, input.get(0));
-            assertThat(getResult).isEqualTo(compare.get(0));
-            assertThat(GsonWrapper.fromJson(getResult, OperationData.class)).isEqualTo(GsonWrapper.fromJson(compare.get(0), OperationData.class));
         };
     }
 
@@ -184,15 +166,20 @@ public final class OperationContractTest extends TestCreationBase {
             String operationJson = "";
             for (String id: ids) {
                 Context ctx = TestUtil.mockContext(stub, id);
-                operationJson = contract.approveTransaction(ctx, "", MatriculationDataContract.contractName, "addMatriculationData", GsonWrapper.toJson(input));
+                operationJson = contract.initiateOperation(ctx, "", MatriculationDataContract.contractName, "addMatriculationData", GsonWrapper.toJson(input));
             }
             Context ctx = TestUtil.mockContext(stub);
             MatriculationDataContract matriculationContract = new MatriculationDataContract();
             stub.setFunction(MatriculationDataContract.contractName + ":addMatriculationData");
             matriculationContract.addMatriculationData(ctx, input.get(0));
+            matriculationContract.addMatriculationData(ctx, input.get(0));
             String operationId = GsonWrapper.fromJson(operationJson, OperationData.class).getOperationId();
             stub.setFunction("UC4.OperationData:approveTransaction");
-            OperationData operation = GsonWrapper.fromJson(contract.getOperationData(ctx, operationId), OperationData.class);
+            Type listType = new TypeToken<ArrayList<OperationData>>() {
+
+            }.getType();
+            List<OperationData> operations = GsonWrapper.fromJson(contract.getOperations(ctx, GsonWrapper.toJson(Collections.singletonList(operationId)), "", "", "", "", ""), listType);
+            OperationData operation = operations.get(0);
             OperationDataState expectedState = GsonWrapper.fromJson(compare.get(0), OperationDataState.class);
             assertThat(operation.getState()).isEqualTo(expectedState);
         };